This deserializes multiple choice
