+++ client/js/middleware/serialization/qbank/deserializers/factory.js
@@ -1,8 +1,9 @@
 import multipleChoice   from './multiple_choice';
+import genusTypes       from '../../../../constants/genus_types';
 
 export default function factory(type) {
   switch (type) {
-    case 'multipleChoice':
+    case genusTypes.item.multipleChoice:
       return multipleChoice;
 
     default:
